put code on RedLightExcitedVisitorAgent in a separate file
